# get_started_documentation/openapi.json
{
  "info": {
    "title": "Assistant Skills Service API",
    "version": "1.0"
  },
  "openapi": "3.0.3",
  "servers": [
    {
      "url": "https://assistant.us-east-2.intelligence.webex.com/skills/api"
    }
  ],
  "paths": {
    "/developers/{user_id}/skills": {
      "post": {
        "tags": [
          "Developer Skills"
        ],
        "summary": "Creates a skill",
        "description": "Creates a skill using the information provided. This will also add a developer registrationto the skill, so the developer can start using it right after creation.",
        "responses": {
          "200": {
            "description": "The newly created skill.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillCreateResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "description": "id of developer",
            "required": true,
            "allowEmptyValue": false,
            "deprecated": false,
            "in": "path",
            "name": "user_id"
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SkillCreateRequest"
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Developer Skills"
        ],
        "summary": "Gets all skills",
        "description": "Returns the skills owned by the developer.",
        "responses": {
          "200": {
            "description": "The developer skills.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SkillGetResponse"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "description": "id of developer",
            "required": true,
            "allowEmptyValue": false,
            "deprecated": false,
            "in": "path",
            "name": "user_id"
          }
        ]
      }
    },
    "/developers/{user_id}/skills/{skill_id}": {
      "get": {
        "tags": [
          "Developer Skills"
        ],
        "summary": "Gets a skill",
        "description": "Returns the skill specified by the id in the path.",
        "responses": {
          "200": {
            "description": "The skill requested.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillGetResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "description": "id of developer",
            "required": true,
            "allowEmptyValue": false,
            "deprecated": false,
            "in": "path",
            "name": "user_id"
          },
          {
            "schema": {
              "type": "string"
            },
            "description": "id of skill",
            "required": true,
            "allowEmptyValue": false,
            "deprecated": false,
            "in": "path",
            "name": "skill_id"
          }
        ]
      },
      "patch": {
        "tags": [
          "Developer Skills"
        ],
        "summary": "Modifies a skill",
        "description": "Updates an existing skill with the information provided.",
        "responses": {
          "200": {
            "description": "The modified skill.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillGetResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "description": "id of developer",
            "required": true,
            "allowEmptyValue": false,
            "deprecated": false,
            "in": "path",
            "name": "user_id"
          },
          {
            "schema": {
              "type": "string"
            },
            "description": "id of skill",
            "required": true,
            "allowEmptyValue": false,
            "deprecated": false,
            "in": "path",
            "name": "skill_id"
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SkillUpdateRequest"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Developer Skills"
        ],
        "summary": "Deletes a skill",
        "description": "This will delete a skill. If the 'HARD_DELETE' query param is specifed and there are no orgregistrations, the skill will be completely deleted from the service. Otherwise this justmakes a soft delete.",
        "responses": {
          "204": {
            "description": "The resource was deleted successfully."
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "description": "id of developer",
            "required": true,
            "allowEmptyValue": false,
            "deprecated": false,
            "in": "path",
            "name": "user_id"
          },
          {
            "schema": {
              "type": "string"
            },
            "description": "id of skill",
            "required": true,
            "allowEmptyValue": false,
            "deprecated": false,
            "in": "path",
            "name": "skill_id"
          },
          {
            "schema": {
              "type": "string"
            },
            "description": "Specifies if we the skills needs to be hard deleted.",
            "required": false,
            "allowEmptyValue": false,
            "deprecated": false,
            "in": "query",
            "name": "HARD_DELETE"
          }
        ]
      }
    },
    "/developers/{user_id}/registrations": {
      "get": {
        "tags": [
          "Developer Registrations"
        ],
        "summary": "Gets all developer registrations",
        "description": "Returns the developer registrations.",
        "responses": {
          "200": {
            "description": "The developer registrations.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeveloperRegistrationGetResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "description": "id of developer",
            "required": true,
            "allowEmptyValue": false,
            "deprecated": false,
            "in": "path",
            "name": "user_id"
          }
        ]
      }
    },
    "/developers/{user_id}/registrations/{registration_id}": {
      "get": {
        "tags": [
          "Developer Registrations"
        ],
        "summary": "Gets a developer registration",
        "description": "Returns the developer registration specified by the id in the path.",
        "responses": {
          "200": {
            "description": "The registrations requested.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeveloperRegistrationGetResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "description": "id of developer",
            "required": true,
            "allowEmptyValue": false,
            "deprecated": false,
            "in": "path",
            "name": "user_id"
          },
          {
            "schema": {
              "type": "string"
            },
            "description": "id of the registration",
            "required": true,
            "allowEmptyValue": false,
            "deprecated": false,
            "in": "path",
            "name": "registration_id"
          }
        ]
      },
      "patch": {
        "tags": [
          "Developer Registrations"
        ],
        "summary": "Modifies a developer registration.",
        "description": "Updates an existing developer registration with the information provided.",
        "responses": {
          "200": {
            "description": "The modified skill.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeveloperRegistrationUpdateResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "description": "id of developer",
            "required": true,
            "allowEmptyValue": false,
            "deprecated": false,
            "in": "path",
            "name": "user_id"
          },
          {
            "schema": {
              "type": "string"
            },
            "description": "id of the registration",
            "required": true,
            "allowEmptyValue": false,
            "deprecated": false,
            "in": "path",
            "name": "registration_id"
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeveloperRegistrationUpdateRequest"
              }
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "BearerAuthorization": []
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuthorization": {
        "scheme": "bearer",
        "type": "http",
        "description": "Token issued from Common Identity"
      }
    },
    "schemas": {
      "ISOLanguageTypes": {
        "title": "ISOLanguageTypes",
        "description": "Languages supported for skills.",
        "enum": [
          "en",
          "es",
          "fr"
        ]
      },
      "SkillCreateRequest": {
        "title": "SkillCreateRequest",
        "type": "object",
        "properties": {
          "languages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ISOLanguageTypes"
            }
          },
          "url": {
            "title": "Url",
            "minLength": 1,
            "maxLength": 65536,
            "format": "uri",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "contact_email": {
            "title": "Contact Email",
            "type": "string",
            "format": "email"
          },
          "secret": {
            "title": "Secret",
            "type": "string"
          },
          "public_key": {
            "title": "Public Key",
            "type": "string"
          },
          "suggested_invocation_names": {
            "title": "Suggested Invocation Names",
            "default": [],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "languages",
          "url",
          "name",
          "contact_email",
          "secret",
          "public_key"
        ]
      },
      "SkillCreateResponse": {
        "title": "SkillCreateResponse",
        "type": "object",
        "properties": {
          "skill_id": {
            "title": "Skill Id",
            "type": "string",
            "format": "uuid"
          },
          "developer_id": {
            "title": "Developer Id",
            "type": "string",
            "format": "uuid"
          },
          "url": {
            "title": "Url",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "contact_email": {
            "title": "Contact Email",
            "type": "string"
          },
          "public": {
            "title": "Public",
            "type": "boolean"
          },
          "deleted": {
            "title": "Deleted",
            "type": "boolean"
          },
          "created_at": {
            "title": "Created At",
            "type": "string",
            "format": "date-time"
          },
          "deleted_at": {
            "title": "Deleted At",
            "type": "string",
            "format": "date-time"
          },
          "modified_at": {
            "title": "Modified At",
            "type": "string",
            "format": "date-time"
          },
          "suggested_invocation_names": {
            "title": "Suggested Invocation Names",
            "default": [],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "languages": {
            "title": "Languages",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "skill_id",
          "developer_id",
          "url",
          "name",
          "contact_email",
          "public",
          "deleted",
          "created_at",
          "modified_at",
          "languages"
        ]
      },
      "SkillGetResponse": {
        "title": "SkillGetResponse",
        "type": "object",
        "properties": {
          "skill_id": {
            "title": "Skill Id",
            "type": "string",
            "format": "uuid"
          },
          "developer_id": {
            "title": "Developer Id",
            "type": "string",
            "format": "uuid"
          },
          "url": {
            "title": "Url",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "contact_email": {
            "title": "Contact Email",
            "type": "string"
          },
          "public": {
            "title": "Public",
            "type": "boolean"
          },
          "deleted": {
            "title": "Deleted",
            "type": "boolean"
          },
          "created_at": {
            "title": "Created At",
            "type": "string",
            "format": "date-time"
          },
          "deleted_at": {
            "title": "Deleted At",
            "type": "string",
            "format": "date-time"
          },
          "modified_at": {
            "title": "Modified At",
            "type": "string",
            "format": "date-time"
          },
          "suggested_invocation_names": {
            "title": "Suggested Invocation Names",
            "default": [],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "languages": {
            "title": "Languages",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "skill_id",
          "developer_id",
          "url",
          "name",
          "contact_email",
          "public",
          "deleted",
          "created_at",
          "modified_at",
          "languages"
        ]
      },
      "SkillUpdateRequest": {
        "title": "SkillUpdateRequest",
        "type": "object",
        "properties": {
          "languages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ISOLanguageTypes"
            }
          },
          "url": {
            "title": "Url",
            "minLength": 1,
            "maxLength": 65536,
            "format": "uri",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "contact_email": {
            "title": "Contact Email",
            "type": "string",
            "format": "email"
          },
          "secret": {
            "title": "Secret",
            "type": "string"
          },
          "public_key": {
            "title": "Public Key",
            "type": "string"
          },
          "suggested_invocation_names": {
            "title": "Suggested Invocation Names",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "DeveloperRegistrationGetResponse": {
        "title": "DeveloperRegistrationGetResponse",
        "type": "object",
        "properties": {
          "registration_id": {
            "title": "Registration Id",
            "type": "string",
            "format": "uuid"
          },
          "skill_id": {
            "title": "Skill Id",
            "type": "string",
            "format": "uuid"
          },
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "languages": {
            "title": "Languages",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "invocation_names": {
            "title": "Invocation Names",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "created_at": {
            "title": "Created At",
            "type": "string",
            "format": "date-time"
          },
          "modified_at": {
            "title": "Modified At",
            "type": "string",
            "format": "date-time"
          },
          "developer_id": {
            "title": "Developer Id",
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "registration_id",
          "skill_id",
          "enabled",
          "languages",
          "invocation_names",
          "created_at",
          "modified_at",
          "developer_id"
        ]
      },
      "DeveloperRegistrationUpdateRequest": {
        "title": "DeveloperRegistrationUpdateRequest",
        "type": "object",
        "properties": {
          "invocation_names": {
            "title": "Invocation Names",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "DeveloperRegistrationUpdateResponse": {
        "title": "DeveloperRegistrationUpdateResponse",
        "type": "object",
        "properties": {
          "registration_id": {
            "title": "Registration Id",
            "type": "string",
            "format": "uuid"
          },
          "skill_id": {
            "title": "Skill Id",
            "type": "string",
            "format": "uuid"
          },
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "languages": {
            "title": "Languages",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "invocation_names": {
            "title": "Invocation Names",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "created_at": {
            "title": "Created At",
            "type": "string",
            "format": "date-time"
          },
          "modified_at": {
            "title": "Modified At",
            "type": "string",
            "format": "date-time"
          },
          "developer_id": {
            "title": "Developer Id",
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "registration_id",
          "skill_id",
          "enabled",
          "languages",
          "invocation_names",
          "created_at",
          "modified_at",
          "developer_id"
        ]
      }
    }
  }
}